+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 🚀 Features
 
+- Export token info to Multichain service ([#12867](https://github.com/blockscout/blockscout/pull/12867))
 - Export balances to Multichain DB([#12726](https://github.com/blockscout/blockscout/pull/12726))
 - Add eip7702 authorization status fetcher ([#12451](https://github.com/blockscout/blockscout/issues/12451))
 - Add token1155tx token404tx api v1 endpoints ([#12720](https://github.com/blockscout/blockscout/issues/12720))